Update tests for version 3
Split tests in smaller files and ensure that a name is always provided.
